[#214] Move common upload logic into abstract class
